Adding working base router using dependency injection
