+++ growintandem-web-service/src/main/kotlin/varilla/joseph/growintandem/http/HttpRouter.kt
@@ -1,8 +1,13 @@
 package varilla.joseph.growintandem.http
 
+import io.vertx.ext.web.Router
 import org.koin.core.KoinComponent
 
 interface HttpRouter :KoinComponent {
+  /**
+   * Returns the router
+   */
+  suspend fun getRouter(): Router
 
   /**
    * Returns all the plants in the system
