+++ growintandem-web-service/src/main/kotlin/varilla/joseph/growintandem/modules/application/VertxModules.kt
@@ -1,0 +1,9 @@
+package varilla.joseph.growintandem.modules.application
+
+
+import io.vertx.core.Vertx
+import org.koin.dsl.module
+
+val VertxModule = module {
+  single { Vertx.vertx() as Vertx }
+}
