+++ growintandem-web-service/src/main/kotlin/varilla/joseph/growintandem/modules/http/HttpRouterModules.kt
@@ -1,0 +1,9 @@
+package varilla.joseph.growintandem.modules.http
+
+import org.koin.dsl.module
+import varilla.joseph.growintandem.http.HttpRouter
+import varilla.joseph.growintandem.http.HttpRouterImpl
+
+val HttpRouterModules = module {
+  single { HttpRouterImpl(get()) as HttpRouter }
+}
